Release merge to master

+++ Compiler.py
@@ -1,38 +1,45 @@
 import shutil
-import PyInstaller.__main__
+import PyInstaller.__main__ as pyinstaller
 import os
-import inquirer
-import inquirer.themes
-import termcolor
+import KDS.System
 from datetime import datetime
 
-AppDataPath = os.path.join(os.getenv('APPDATA'), "KL Corporation", "KDS Compiler")
+AppDataPath = os.path.join(str(os.getenv('APPDATA')), "KL Corporation", "KDS Compiler")
 BuildsPath = os.path.join(AppDataPath, "Builds")
 os.makedirs(BuildsPath, exist_ok=True)
 WorkPath = os.path.join(AppDataPath, "cache", "work")
 CachePath = os.path.join(AppDataPath, "cache")
 if os.path.isdir(CachePath): shutil.rmtree(CachePath)
 
-questions = [
-    inquirer.Checkbox("toBuild", "Which KDS Modules you want to build?", ("Koponen Dating Simulator", "Level Builder"), "Koponen Dating Simulator")
-]
-
-answers = inquirer.prompt(questions)
-
 parentDir = os.path.dirname(os.path.abspath(__file__))
 
 BuildPath = os.path.join(BuildsPath, "build_" + datetime.now().strftime("%Y-%m-%d-%H-%M-%S"))
 EditorTexturesPath = os.path.join(BuildPath, "KoponenDatingSimulator", "Assets", "Textures", "Editor")
-BuiltNames = []
-for buildType in answers["toBuild"]:
-    if buildType == "Koponen Dating Simulator":
-        fileName = "KoponenDatingSimulator.py"
-        iconName = "gameIcon.ico"
-    else:
-        fileName = "LevelBuilder.py"
-        iconName = "levelBuilderIcon.ico"
-        
-    PyInstaller.__main__.run([
+toBuild = {
+    "Koponen Dating Simulator": {
+        "filename": "KoponenDatingSimulator.py",
+        "iconname": "gameIcon.ico",
+        "keepEditor": False
+    },
+    "Level Builder": {
+        "filename": "LevelBuilder.py",
+        "iconname": "levelBuilderIcon.ico",
+        "keepEditor": True
+    }
+}
+
+def clearCache():
+    if os.path.isdir(CachePath):
+        print("Clearing cache...")
+        shutil.rmtree(CachePath)
+
+for buildType in toBuild.items():
+    clearCache()
+
+    filename = buildType[1]["filename"]
+    iconname = buildType[1]["iconname"]
+
+    pyinstaller.run([
         "--noconfirm",
         "--distpath",
         BuildPath,
@@ -42,22 +49,18 @@ for buildType in answers["toBuild"]:
         CachePath,
         "--windowed",
         "--icon",
-        f"{parentDir}/Assets/Textures/Branding/{iconName}",
+        f"{parentDir}/Assets/Textures/Branding/{iconname}",
         "--add-data",
         f"{parentDir}/Assets;Assets/",
         "--paths",
         f"{parentDir}/KDS",
-        f"{parentDir}/{fileName}"
+        f"{parentDir}/{filename}"
     ])
-    if os.path.isdir(EditorTexturesPath):
+    if not buildType[1]["keepEditor"] and os.path.isdir(EditorTexturesPath):
         shutil.rmtree(EditorTexturesPath)
         print(f"Deleted Editor Textures directory at {EditorTexturesPath}")
-    BuiltNames.append((buildType, os.path.splitext(fileName)[0]))
 
-if len(answers["toBuild"]) < 1:
-    print(termcolor.colored("Nothing was built.", "red"))
-else:
-    for name in BuiltNames:
-        print(termcolor.colored(f"Built {name[0]} at " + os.path.join(BuildPath, name[1], name[1] + ".exe"), "green"))
-
-if os.path.isdir(os.path.join(AppDataPath, "cache")): shutil.rmtree(os.path.join(AppDataPath, "cache"))
+clearCache()
+for buildType in toBuild.items():
+    cutName = os.path.splitext(buildType[1]["filename"])[0]
+    print(KDS.System.Console.Colored(f"Built {buildType[0]} at " + os.path.join(BuildPath, cutName, f"{cutName}.exe"), "green"))
